+++ carp_mobile_sensing/lib/core/study.dart
@@ -140,7 +140,7 @@ class FileDataEndPoint extends DataEndPoint {
   Map<String, dynamic> toJson() => _$FileDataEndPointToJson(this);
 
   String toString() =>
-      'FILE - buffer ${(bufferSize / 1000).round()} KB${zip ? ', zipped' : ''}, ${encrypt ? ', encrypted' : ''}';
+      'FILE - buffer ${(bufferSize / 1000).round()} KB${zip ? ', zipped' : ''}${encrypt ? ', encrypted' : ''}';
 }
 
 /// A enumeration of known endpoint API types.