+++ carp_mobile_sensing_app/lib/src/blocs/sensing_bloc.dart
@@ -13,7 +13,7 @@ class SensingBloc {
   void init() async {}
 
   void start() async {
-    sensing.start();
+    await sensing.start();
   }
 
   void pause() {