+++ carp_mobile_sensing_app/lib/src/sensing/sensing.dart
@@ -21,7 +21,7 @@ class Sensing {
   }
 
   /// Start sensing.
-  void start() async {
+  Future<void> start() async {
     // Get the study.
     study = await mock.getStudy(testStudyId);